Update getback.c
Provides more precision in distances > 1km and larger font for the distance.

Hi, Samuel.  This is one of my favorite apps.  I don't want to take it over, but I have a few ideas and I'm trying to implement them as a way of learning to program Pebble C.  

Cheers,

Peter